Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -3,7 +3,7 @@
 
 
 int main() {
-    int a,b
+    float a,b
     scanf("%f",&a);
     scanf("%f",&b);
     printf("Product: %f",a * b);
